rename project, update licensing
