+++ LICENSE
@@ -1,17 +1,21 @@
-The project as a whole is licensed under the GNU General Public License v3.0 (GPLv3).
+MIT License
 
-See LICENSE.GPL for the full text of the GPLv3 License.
+Copyright (c) 2026 Manhhao
 
-EXCEPTION / DUAL LICENSE:
-A part of this source code is also available under the MIT License.
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
 
-Files subject to GPLv3:
-- src/deinflector.cpp
-- src/preprocessor/text_processor.cpp
-- include/yomitandicts/deinflector.hpp
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
 
-If you wish to use parts of the project under the MIT License, you must:
-1. Remove the GPLv3 restricted files listed above.
-2. Not link against or include the GPLv3 restricted files in your project.
-
-See LICENSE.MIT for the full text of the MIT License.
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
