+++ include/hoshidicts.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "hoshidicts/deconjugator.hpp"
+#include "hoshidicts/importer.hpp"
+#include "hoshidicts/lookup.hpp"
+#include "hoshidicts/query.hpp"
