+++ include/hoshidicts/importer.hpp
@@ -8,7 +8,6 @@ struct ImportResult {
   std::string title;
   size_t term_count = 0;
   size_t meta_count = 0;
-  size_t tag_count = 0;
   size_t media_count = 0;
   std::vector<std::string> errors;
 };
