+++ include/hoshidicts/lookup.hpp
