+++ include/hoshidicts/query.hpp
