--- include/yomitandicts.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "yomitandicts/deconjugator.hpp"
-#include "yomitandicts/importer.hpp"
-#include "yomitandicts/lookup.hpp"
-#include "yomitandicts/query.hpp"
